refactor(web): apply changed paths to import components
