--- apps/web/src/widgets/modal/model/constants/index.ts
@@ -1,4 +1,0 @@
-export * from './delete';
-export * from './vote';
-export * from './template';
-export * from './schedule';
