--- apps/web/src/widgets/modal/model/index.ts
@@ -1,1 +1,0 @@
-export * from './constants';
