--- apps/web/src/widgets/modal/types/vote.ts
@@ -1,3 +1,0 @@
-import { VOTE } from '~/widgets/modal/model';
-
-export type VoteSettingOption = keyof typeof VOTE;
