--- apps/web/src/widgets/modal/ui/member/index.ts
@@ -1,3 +1,0 @@
-export { default as ProjectMemberManageModal } from './ProjectMemberManageModal';
-export { default as LevelInfoModal } from './LevelInfoModal';
-export * from './LevelTable';
